Avoid latent buffer overrun in win32read.
Mostly ignore the 'length' parameter of win32read instead of using it
in a wrong and dangerous way that creates a latent buffer overrun bug
when calling ReadConsoleInput.

(The argument nLength of ReadConsoleInput is defined as:
"The size of the array pointed to by the lpBuffer parameter, in array elements.")

The bug was latent as win32read is currently only called with length == 1.

The rest of win32read does not handle more than one event right now, anyway.